MultiExchangeBot/requirements.txt
python-dotenv>=1.0.0
aiohttp>=3.9.0
websockets>=12.0
fastapi>=0.100.0
uvicorn>=0.22.0
